requirements-dev.txt
wheel==0.38.1
coverage==7.2.7
setuptools==67.8.0;python_version>="3.12"
requests==2.31.0

# Pytest specific deps
pytest==7.4.0
pytest-cov==4.1.0
atomicwrites>=1.0 # Windows requirement
colorama>0.3.0 # Windows requirement
python-dotenv==1.0.0
black==23.11.0
sphinx==5.3.0
furo==2022.12.7
sphinx_copybutton==0.5.1
sphinx-remove-toctrees==0.0.3
tuspy==0.2.5